+++ packages/crayons-extended/custom-objects/src/constants/article-links.ts
@@ -1,0 +1,2 @@
+export const DYNAMIC_SECTION =
+  'https://crmsupport.freshworks.com/en/support/solutions/articles/50000010284';